python_requirements.txt
kafka-python==1.3.5
smart-open==1.5.3